requirements.txt
beautifulsoup4==4.8.2
Flask==2.1.0
Flask-Bootstrap==3.3.7.1
Werkzeug==2.2.2
gunicorn==20.0.4
jsonschema==3.2.0
lxml==4.6.3
numpy~=1.19.5
pandas==1.2.5
pymongo==3.8.0
python-dateutil~=2.8.1
python-keystoneclient==4.0.0
python-swiftclient==3.9.0
redis==3.5.3
regex==2017.4.5
requests==2.19.1
retry~=0.9.2
rq==1.9.0
Unidecode==1.0.22
